replace gene_range_time_limit with single_recurse_max_num_ops
which makes the output fully deterministic.

In case the whole region takes a long time, recurse_region_max_num_ops is a second limit, extending the use of num_ops to track multiple recursive calls

Also change num_ops to be more granular so that it tracks more closely with elapsed time.